requirements.txt
apache-airflow[celery]==2.4.1
psycopg2-binary==2.9.4
kafka-python==2.0.2
